requirements.txt
flask>=2.3,<4.0
opencv-python>=4.8,<5.0
face-recognition>=1.3,<2.0
numpy>=1.24,<2.0
pandas>=2.0,<3.0
mysql-connector-python>=8.1,<9.0
bcrypt>=4.0,<5.0
Pillow>=10.0,<11.0
